requirements.txt
Jinja2==3.1.4
livereload==2.7.0
python-frontmatter==1.1.0
PyYAML==6.0.2
